ml-service/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Optional: Deep Learning (jika menggunakan Neural Networks)
# tensorflow==2.15.0
# keras==2.15.0

# Data Validation
pydantic==2.5.2

# HTTP Requests (untuk komunikasi dengan backend)
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Model Persistence
joblib==1.3.2

# API Documentation (Optional)
flasgger==0.9.7.1

# Production Server
gunicorn==21.2.0

# Data Science & Visualization (for Jupyter development)
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Kaggle API (for dataset download)
kaggle==1.5.16

# Additional utilities
tqdm==4.66.1  # Progress bars
openpyxl==3.1.2  # Excel support
